Fix how unedited files is handled in Buffer (save edit history correctly).
This change makes the code much smaller and simpler.  This is because
it reduces an edge case in every function manipulating the
StateNodes.  Instead of having to check if the state is None and then
doing a separate set of behavior, it is never None and instead does
nothing when initialized.

Also added tests to ensure this behavior is satisfied.

>
Hello Worl
>Append d
Hello World
>Undo
Hello Worl
>Redo
Hello World