Refactor `CursesRenderer` into separate crate: `ted_renderers`.
This also replaces the backend of `CursesRenderer` with
`pancurses-result`.  This is a crate I wrote that aims to safely wrap
`pancurses`.  Thus we no longer need to implement `Send` and `Sync`
for `CursesRenderer` manually because it is done automatically.

The destructor also has to be refactored.  Now to print the log after
shell mode is restored (as curses has been destructed), we just add a
marker object that calls `print_log` when it is dropped.